read calibration section also when ignore_flags=T
So far the 'calibration' section of the config file was only interpreted when 'ignore_flags=False' was handed to the 'config()' function AND 'calibrate_flag: True' was in the config file.
The information of the 'calibration' section is required in 'pyfocs_ex1_rawdatacheck.ipynb' and 'pyfocs_ex2_calibcheck.ipynb'.
And while 'ignore_flags' could be turned to 'False' in the notebooks, the 'calibrate_flag' in the config file still would have to be 'True', which is inconvenient for the first notebook, after just performing the archiving and archive-read step.
Therefore a section was added (line 120 to 134), interpreting the 'calibration' in case 'ignore_flags=True'.
Here, the only check performed is, whether each calibration fiber-section is referenced with two valid LAF statements.
This code will also be executed in 'pyfocs_ex2_calibcheck.ipynb'. However, at this point the complete check of the 'calibration' section was already done during calibration, so there is no need to run through the whole check in the notebook.